backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.32
psycopg2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.4.0
pydantic[email]
python-jose[cryptography]==3.3.0
python-multipart==0.0.9